refactor clean up
